Split Config into ConfigManager and Config Sections
